[fix] 경로및 버그 수정

+++ core/linux/bin/graphics/install_drawing.sh
@@ -41,9 +41,9 @@ Icon=${ICON_PATH}
 Categories=${APP_GRP}
 Terminal=false"
 
-    if [[ *"${DESKTOP_PATH}"* == *".local"* ]]; then
+    if [[ *"${DESKTOP_PATH}"* == *"\/home"* ]]; then
         # ~/.local/share/applications/com.github.maoschanz.drawing.desktop
-        su - ${CUR_USER} -c "echo '${DESKTOP_CMD}' > ${DESKTOP_PATH}";
+        su - ${CUR_USER} -c "echo \"${DESKTOP_CMD}\" > ${DESKTOP_PATH}";
     else
         # /usr/share/applications/com.github.maoschanz.drawing.desktop
         echo "${DESKTOP_CMD}" > ${DESKTOP_PATH};
@@ -115,7 +115,9 @@ function install_drawing_for_nix()
         fi
         # ----------------------------------------------------------------------
 
-        su - ${CUR_USER} -c "ln -s ${NIX_DESKTOP_PATH} ${DESKTOP_PATH}";
+        if [[ ! -f ${DESKTOP_PATH} ]]; then
+            su - ${CUR_USER} -c "ln -s ${NIX_DESKTOP_PATH} ${DESKTOP_PATH}";
+        fi
     else
         set_desktop;
     fi
